culinary_heaven/requirements/prod.txt
Django==5.0.1
django-environ==0.11.2
django-filter==24.3
django-mptt==0.16.0
django-select2==8.2.0
django-simple-captcha==0.6.0
django-star-ratings==0.9.2
django-user-agents==0.4.0
pillow==10.2.0
pytils==0.4.1
transliterate==1.10.2
django-redis==5.4.0
gunicorn==20.1.0
psycopg2-binary==2.9.9